fix(__dirname): adds horrible wrapper hack for __dirname with webpack & electron & workers
This is an abominable hack which works around the fact webpack just does whatever the fuck it wants
with workers. We stall everything in the worker until we've sent it the correct path to the .asar,
then in the NDI store module we temporarily set the location of grandiose within the .asar.
Guaranteed this will come back to bite us in the arse.

fix #114